Eliminate deadlock opening channel when connection is closing.
If we try to open a channel as a connection is being closed, we're
seeing cases where channels get deadlocked trying to open. We played
with a few different solutions, but simply having a 'closed' variable on
the connection which is set on shutdown (under lock) and using this
variable (under lock) before allocating the channel to make sure the
connection is still valid.

With this check, a channel will either be rejected because the
connection is closed/closing, or the channel will be added to the
connections list of channels before shutdown closes all channels.